% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissevote.gen.R
\name{ballot_dates_federal}
\alias{ballot_dates_federal}
\title{Get federal ballot dates}
\usage{
ballot_dates_federal(
  source = c("Neuchatel", "Zurich"),
  exclude_counterproposals = TRUE,
  use_cache = TRUE,
  max_cache_age = "1 day"
)
}
\arguments{
\item{source}{Federal ballot dates can either be retrieved from the official website of the canton of \emph{Neuchatel} (default) or \emph{Zurich}. Note that only
referendum dates will be returned from \code{source = "Zurich"}, election dates can only be scraped from \code{source = "Neuchatel"}. On the other hand,
\code{source = "Zurich"} will return referendum dates back to the 19th century, while \code{source = "Neuchatel"} only goes back to the turn of the millennium.}

\item{exclude_counterproposals}{Exclude indirect counterproposals and tie-breaker questions. A logical scalar.}

\item{use_cache}{Whether or not to return cached results if possible. If \code{FALSE}, results are always newly fetched regardless of \code{max_cache_age}.}

\item{max_cache_age}{Duration after which cached results are refreshed (i.e. newly fetched). A valid \link[lubridate:as.duration]{lubridate duration}. Use \code{Inf} to disable cache expiry. Only relevant if \code{use_cache = TRUE}.}
}
\value{
\code{pkgsnippets::return_label("data")}
}
\description{
Get federal ballot dates
}
