% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissevote.gen.R
\name{ballot_dates_municipal}
\alias{ballot_dates_municipal}
\title{Get municipal ballot dates}
\usage{
ballot_dates_municipal(
  cantons = c("Geneva", "Neuchatel", "Zurich"),
  exclude_counterproposals = TRUE,
  use_cache = TRUE,
  cache_lifespan = "1 day"
)
}
\arguments{
\item{cantons}{The names of the cantons to retrieve ballot dates for. A character vector.}

\item{exclude_counterproposals}{Exclude indirect counterproposals and tie-breaker questions. A logical scalar.}

\item{use_cache}{Return cached results if possible. If \code{FALSE}, results are always newly fetched regardless of \code{cache_lifespan}.}

\item{cache_lifespan}{The duration after which cached results are refreshed (i.e. newly fetched). A valid \link[lubridate:as.duration]{lubridate duration}. Only relevant if \code{use_cache = TRUE}.}
}
\value{
\code{pkgsnippets::return_label("data")}
}
\description{
Get municipal ballot dates
}
