% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissevote.gen.R
\name{ballot_dates_raw_data}
\alias{ballot_dates_raw_data}
\title{List ballot dates for which raw data files are available}
\usage{
ballot_dates_raw_data(canton = c("Bern", "Geneva", "Neuchatel"))
}
\arguments{
\item{canton}{The canton for which to list ballot dates. A character scalar.}
}
\value{
A vector of \link[base:Dates]{dates}.
}
\description{
List ballot dates for which raw data files are available
}
