% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissevote-GEN.R
\name{convert_canton_names}
\alias{convert_canton_names}
\title{Convert between full canton names and their two-letter abbreviations}
\usage{
convert_canton_names(cantons)
}
\arguments{
\item{cantons}{The full names or two-letter abbreviations of the cantons to be converted. A character vector.}
}
\value{
A character vector.
}
\description{
Convert between full canton names and their two-letter abbreviations
}
\examples{
convert_canton_names(c("ZH", "VD", "AG"))
convert_canton_names(c("Zurich", "Vaud", "Argovia"))
convert_canton_names(c("ZH", "Vaud", "Argovia"))
}
