% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissevote.gen.R
\name{determine_3_code_combo}
\alias{determine_3_code_combo}
\title{Determine combination of three-code variables (\code{combined_election_procedure})}
\usage{
determine_3_code_combo(...)
}
\arguments{
\item{...}{Up to three individual codes as integers. \code{pkgsnip::roxy_lbl("dyn_dots_support")}}
}
\value{
An integer scalar.
}
\description{
Determine combination of three-code variables (\code{combined_election_procedure})
}
\keyword{internal}
