% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissevote-GEN.R
\name{determine_7_code_combo}
\alias{determine_7_code_combo}
\title{Determine combination of seven-code variables (like \code{combined_level})}
\usage{
determine_7_code_combo(...)
}
\arguments{
\item{...}{Up to three individual codes as integers. \code{pkgsnip::roxy_label("dyn_dots_support")}}
}
\value{
An integer scalar.
}
\description{
Determine combination of seven-code variables (like \code{combined_level})
}
