% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissevote-GEN.R
\name{merge_municipalities}
\alias{merge_municipalities}
\title{Retroactively merge municipalities that have merged}
\usage{
merge_municipalities(data)
}
\arguments{
\item{data}{The data in which to perform the retroactive municipality mergers.}
}
\value{
\code{pkgsnip::return_label("data")}
}
\description{
This function builds upon \code{\link[swissmuni:mutations]{swissmuni::mutations()}} to find municipalities in \code{data} that have merged into another municipality and synthetically applies
this merger "back in time" as best as possible. This allows subsequent data analysis to better track/compare the development of e-voting usage over time.
}
