% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissevote-GEN.R
\name{read_raw_data}
\alias{read_raw_data}
\title{Read in raw data files of a specific Swiss canton}
\usage{
read_raw_data(
  canton = c("Bern", "Geneva", "Neuchatel"),
  ballot_date,
  specific_datasets = NULL
)
}
\arguments{
\item{canton}{The canton of which to read in raw data files.}

\item{ballot_date}{The ballot date of which to read in raw data files.}

\item{specific_datasets}{If not \code{NULL}, only the specified filenames will be processed. A list with vote dates as keys and character vectors of specific
filenames (without the filetype extension) as values. Setting \code{specific_datasets} is only sensible in case of \code{canton = "Geneva"} and thus ignored
otherwise.}
}
\value{
In case of \code{canton = "Geneva"}, a list with a tibble for each raw data file plus a metadata tibble. Otherwise a single tibble.
}
\description{
Read in raw data files of a specific Swiss canton
}
