% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissevote-GEN.R
\name{read_raw_data_geneva}
\alias{read_raw_data_geneva}
\title{Read in raw data files from the canton of Geneva}
\usage{
read_raw_data_geneva(
  ballot_date,
  specific_datasets = NULL,
  check_municipalities = TRUE,
  remove_duplicated_votes = TRUE,
  remove_unknown_municipality_district_combos = FALSE,
  remove_NA_locality_codes = FALSE,
  skip_problematic_dates = FALSE
)
}
\arguments{
\item{ballot_date}{The ballot date of which to read in raw data files.}

\item{specific_datasets}{If not \code{NULL}, only the specified filenames will be processed. A list with ballot dates as keys and character vectors of specific
filenames (without the filetype extension) as values.}

\item{check_municipalities}{Check if \code{municipality} could be determined for all entries excluding Swiss living abroad. A logical scalar.}

\item{remove_duplicated_votes}{Remove duplicate votes (where \code{votcorr == "A"}; only if the necessary variable exists). The variable \code{votcorr} is used to
identify voters for whom several voting cards have been issued (due to non-receipt, loss, etc.). The cases with \code{votcorr = "A"} have to be filtered out in
order to get correct results. A logical scalar.}

\item{remove_unknown_municipality_district_combos}{Ignore rows in raw data that are \emph{not} Swiss abroad but missing a \code{municipality} (usually with the
undocumented district code \code{99}). A logical scalar.}

\item{remove_NA_locality_codes}{Ignore rows in raw data that have no \code{municipality_code_2_digits} set (\code{NA}). A logical scalar.}

\item{skip_problematic_dates}{Skip \code{ballot_date}s for which unresolved parsing problems exist. A logical scalar.}
}
\value{
A list with a tibble of each raw data file plus metadata tibble.
}
\description{
Read in raw data files from the canton of Geneva
}
