% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissevote-package.R
\docType{package}
\name{swissevote-package}
\alias{swissevote}
\alias{swissevote-package}
\title{swissevote: Provides functions around the Swiss e-voting trial
    data collected by the Centre for Democracy Studies Aarau (ZDA).}
\description{
swissevote provides functions to parse
    cantonal voting register data, to scrape authority websites for ballot
    data as well as other auxiliary functions around the Swiss e-voting
    trial data collected by the Centre for Democracy Studies Aarau
    (ZDA).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/zdaarau/swissevote}
  \item Report bugs at \url{https://gitlab.com/zdaarau/swissevote/issues}
}

}
\author{
\strong{Maintainer}: Salim Brüggemann \email{salim-b@pm.me} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID})

}
\keyword{internal}
