% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissevote-GEN.R
\name{vote_dates_federal}
\alias{vote_dates_federal}
\title{Get federal vote dates}
\usage{
vote_dates_federal(
  source = c("Neuchatel", "Zurich"),
  exclude_counterproposals = TRUE,
  use_cache = TRUE,
  cache_lifespan = "1 day"
)
}
\arguments{
\item{source}{Federal vote dates can either be retrieved from the official website of the canton of \emph{Neuchatel} (default) or \emph{Zurich}. Note that only
referendum dates will be returned from \code{source = "Zurich"}, election dates can only be scraped from \code{source = "Neuchatel"}. On the other hand,
\code{source = "Zurich"} will return referendum dates back to the 19th century, while \code{source = "Neuchatel"} only goes back to the turn of the millennium.}

\item{exclude_counterproposals}{Exclude indirect counterproposals and tie-breaker questions. A logical scalar.}

\item{use_cache}{Return cached results if possible. If \code{FALSE}, results are always newly fetched regardless of \code{cache_lifespan}.}

\item{cache_lifespan}{The duration after which cached results are refreshed (i.e. newly fetched). A valid \link[lubridate:as.duration]{lubridate duration}. Only relevant if \code{use_cache = TRUE}.}
}
\value{
\code{pkgsnippets::return_label("data")}
}
\description{
Get federal vote dates
}
