% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissevote.gen.R
\name{read_raw_data}
\alias{read_raw_data}
\title{Read in raw data files of a specific Swiss canton}
\usage{
read_raw_data(
  canton = c("Bern", "Geneva", "Neuchatel"),
  ballot_date,
  specific_datasets = NULL,
  use_cache = TRUE,
  max_cache_age = "7 days"
)
}
\arguments{
\item{canton}{The canton of which to read in raw data files.}

\item{ballot_date}{The ballot date of which to read in raw data files.}

\item{specific_datasets}{If not \code{NULL}, only the specified filenames will be processed. A list with ballot dates as keys and character vectors of specific
filenames (without the filetype extension) as values. Setting \code{specific_datasets} is only sensible in case of \code{canton = "Geneva"} and thus ignored
otherwise.}

\item{use_cache}{Return cached results if possible. If \code{FALSE}, results are always newly fetched regardless of \code{max_cache_age}.}

\item{max_cache_age}{The duration after which cached results are refreshed (i.e. newly fetched). A valid \link[lubridate:as.duration]{lubridate duration}. Use \code{Inf} to disable cache expiry. Only relevant if \code{use_cache = TRUE}.}
}
\value{
In case of \code{canton = "Geneva"}, a list with a tibble for each raw data file plus a metadata tibble. Otherwise a single tibble.
}
\description{
Read in raw data files of a specific Swiss canton
}
