% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissevote-package.R
\docType{package}
\name{swissevote-package}
\alias{swissevote}
\alias{swissevote-package}
\title{swissevote: Provides Functions Around the Swiss E-Voting Trial Data Collected by the Centre for Democracy Studies Aarau (ZDA)}
\description{
swissevote can parse cantonal voting register data, scrape authority websites for ballot data and provides other auxiliary functions around the Swiss e-voting trial data collected by the Centre for Democracy Studies Aarau (ZDA) at the University of Zurich, Switzerland.
}
\section{Package configuration}{


Some of swissevote's functionality is controlled via package-specific global configuration which can either be set via \href{https://rdrr.io/r/base/options.html}{R options} or \href{https://en.wikipedia.org/wiki/Environment_variable}{environment variables} (the former take precedence). This configuration includes:\tabular{llll}{
   \strong{Description} \tab \strong{R option} \tab \strong{Environment variable} \tab \strong{Default value} \cr
   Maximal timespan to preserve the package's \href{https://pkgpins.rpkg.dev/}{pkgpins} cache. Cache entries older than this will be deleted upon package loading. \tab \code{swissevote.global_max_cache_age} \tab \code{R_SWISSEVOTE_GLOBAL_MAX_CACHE_AGE} \tab \code{"30 days"} \cr
   the path to the directory holding the cantonal raw data files (which are not part of this package due to legal restrictions and/or concerns regarding voter privacy); see the section \emph{Raw data files} in the package's \href{https://swissevote.rpkg.dev/#raw-data-files}{README} for further details \tab \code{swissevote.path_raw_data} \tab \code{R_SWISSEVOTE_PATH_RAW_DATA} \tab  \cr
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/zdaarau/rpkgs/swissevote}
  \item Report bugs at \url{https://gitlab.com/zdaarau/rpkgs/swissevote/issues}
}

}
\author{
\strong{Maintainer}: Salim Brüggemann \email{salim-b@pm.me} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID})

}
\keyword{internal}
