% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissevote-package.R
\docType{package}
\name{swissevote-package}
\alias{swissevote}
\alias{swissevote-package}
\title{swissevote: Provides Functions Around the Swiss E-Voting Trial Data Collected
    by the Centre for Democracy Studies Aarau (ZDA)}
\description{
swissevote can parse cantonal voting register data, scrape
    authority websites for ballot data and provides other auxiliary
    functions around the Swiss e-voting trial data collected by the Centre
    for Democracy Studies Aarau (ZDA) at the University of Zurich,
    Switzerland.
}
\section{Package options}{


The following package-specific R options are used to control the package's functionality:\tabular{lll}{
   \strong{name} \tab \strong{description} \tab \strong{set automatically during package load} \cr
   \code{swissevote.path_raw_data} \tab the path to the directory holding the cantonal raw data files (which are not part of this package due to legal restrictions and/or concerns regarding voter privacy); see the package's \href{https://gitlab.com/zdaarau/swissevote#raw-data-files}{README} for more details; only set automatically for user=salim \tab ✅ \cr
   \code{swissevote.global_cache_lifespan} \tab the default cache lifespan for all functions taking a \code{cache_lifespan} argument; defaults to 30 days \tab ✅ \cr
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/zdaarau/rpkgs/swissevote}
  \item Report bugs at \url{https://gitlab.com/zdaarau/rpkgs/swissevote/issues}
}

}
\author{
\strong{Maintainer}: Salim Brüggemann \email{salim-b@pm.me} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID})

}
\keyword{internal}
