% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissevote-GEN.R
\name{vote_dates_municipal}
\alias{vote_dates_municipal}
\title{Get municipal vote dates}
\usage{
vote_dates_municipal(
  cantons = c("Geneva", "Neuchatel", "Zurich"),
  exclude_counterproposals = TRUE,
  use_cache = TRUE,
  cache_lifespan = 86400L
)
}
\arguments{
\item{cantons}{The names of the cantons to retrieve vote dates for. A character vector.}

\item{exclude_counterproposals}{Exclude indirect counterproposals and tie-breaker questions. A logical scalar.}

\item{use_cache}{\code{pkgsnip::param_label("use_cache")}}

\item{cache_lifespan}{\code{pkgsnip::param_label("cache_lifespan")}}
}
\value{
\code{pkgsnippets::return_label("data")}
}
\description{
Get municipal vote dates
}
